Moved methods to Base class and removed unused imports.
