+++ app/src/main/java/org/schabi/newpipe/player/BackgroundPlayer.java
@@ -29,14 +29,11 @@ import android.content.SharedPreferences;
 import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.os.Build;
-import android.os.Handler;
 import android.os.IBinder;
-import android.os.Looper;
 import android.preference.PreferenceManager;
 import android.util.Log;
 import android.view.View;
 import android.widget.RemoteViews;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -59,10 +56,6 @@ import org.schabi.newpipe.player.resolver.MediaSourceTag;
 import org.schabi.newpipe.util.BitmapUtils;
 import org.schabi.newpipe.util.NavigationHelper;
 import org.schabi.newpipe.util.ThemeHelper;
-
-import java.util.Date;
-import java.util.Timer;
-import java.util.TimerTask;
 
 import static org.schabi.newpipe.player.helper.PlayerHelper.getTimeString;
 import static org.schabi.newpipe.util.Localization.assureCorrectAppLanguage;
@@ -331,10 +324,6 @@ public final class BackgroundPlayer extends Service {
         final private AudioPlaybackResolver resolver;
         private int cachedDuration;
         private String cachedDurationString;
-
-        private Timer timer = new Timer();
-        private int timerHour = 0;
-        private int timerMinute = 0;
 
         BasePlayerImpl(Context context) {
             super(context);
@@ -640,49 +629,6 @@ public final class BackgroundPlayer extends Service {
             updateNotificationThumbnail();
             updateNotification(R.drawable.ic_replay_white);
             lockManager.releaseWifiAndCpu();
-        }
-
-        @Override
-        public void setTimer(int hourOfDay, int minute) {
-            long time = ((hourOfDay * 60) + minute) * 60 * 1000;
-            if (time == 0) {
-                Toast.makeText(getApplicationContext(), getString(R.string.timer_select_time_message), Toast.LENGTH_SHORT).show();
-                return;
-            }
-            timer.cancel();
-            timer.purge();
-            timerHour = hourOfDay;
-            timerMinute = minute;
-            timer = new Timer();
-            timer.schedule(new TimerTask() {
-                @Override
-                public void run() {
-                    new Handler(Looper.getMainLooper()).post(() -> {
-                        if (DEBUG) Log.d(TAG, getString(R.string.log_stopping_the_player));
-                        onPause();
-                    });
-
-                }
-            }, time);
-            Date d = new Date();
-            d.setTime(System.currentTimeMillis() + time);
-            Toast.makeText(getApplicationContext(), String.format(getString(R.string.player_stop_message), d.toString()), Toast.LENGTH_LONG).show();
-        }
-
-        @Override
-        public void cancelTimer() {
-            timer.cancel();
-            timer.purge();
-        }
-
-        @Override
-        public int getHourOfDay() {
-            return this.timerHour;
-        }
-
-        @Override
-        public int getMinutes() {
-            return this.timerMinute;
         }
     }
 }
