+++ app/src/main/res/values/strings.xml
@@ -595,9 +595,6 @@
 
     <!--  Timer Constants  -->
     <string name="set_timer">Set Timer</string>
-    <string name="invalid_method_call">Invalid method call for :</string>
-    <string name="error_setting_timer">Some error occurred while setting timer.</string>
     <string name="player_stop_message">Player will stop at %s</string>
-    <string name="log_stopping_the_player">Timer finished: Stopping the player</string>
     <string name="timer_select_time_message">Please select an appropriate time.</string>
 </resources>
